requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
pandas
geopandas
matplotlib
contextily
